Use ansible's service module instead of templating /etc/rc.conf.local
This is more aligned with the ansible way of doing things.
The service.py module provides an OpenBsdService class that uses rcctl(8)
to handle system services.

The playbooks themselves have not been tested yet, but initial ad-hoc
testing using ansible(1) suggests we could do without manually templating
/etc/rc.conf.local. See below.

Since the service to be enabled might be running, or not, we restart it
to ensure it picks up the new config, if any, after enabling it.

Since the service to be disabled might be running, or not, we stop it
after disabling it.

$ fgrep pflog /etc/rc.conf.local
$
$ ansible localhost -m service -a "name=pflogd enabled=yes args='-s 256' state=restarted"
localhost | success >> {
    "changed": true,
    "enabled": true,
    "name": "pflogd",
    "state": "started"
}
$ fgrep pflog /etc/rc.conf.local
pflogd_flags=-s 256
$
$ ansible localhost -m service -a "name=pflogd enabled=no state=stopped"
localhost | success >> {
    "changed": true,
    "enabled": false,
    "name": "pflogd",
    "state": "stopped"
}
$ fgrep pflog /etc/rc.conf.local
pflogd_flags=NO
$