Rejig how we handle local variables
Ansible will give the highest priority to those variables provided at
the CLI via the -e switch.

This switch accepts YAML or JSON formatted files via '-e @FILE'. This
file can be empty.

We can drop an empty local-vars.yml after cloning the repo in
bootstrap.sh, which will be ignored via .gitignore.

The key to proper dict overrides is to set hash_behaviour to 'merge' instead of 'replace'.
This is more convenient for the user, since individual dictionary keys can be overriden
No need to copy the whole dictionary.