Change mirror transport and target
Up until now we were using FTP and a hardcoded EU DNS name for this.

Let's not assume all OpenHRC users are in the EU :clown_face: and
move to a CDN-based name.

I know, we're selling out to Fastly, but I am getting old and I am
fresh out of fucks to give today!

As per `pkg_add(1)` and `installurl(5)`, we should have no issues
moving to HTTPS.